Update print.egor() to rely on 'pillar'
Printing 'egor' objects now simply calls pillar:::print.tbl_df() for
each of the elements of the objects. In order to skip the "header" which
we print in our own way (data level plus the "active" adjective if
applicable) it is necessary to add the method for tbl_sum() as printing
of tibbles cannot be controled with arguments nor options.

Package 'pillar' is now imported.

Upgrade RoxygenNote in DESCRIPTION.

References #80.